fixed create movie (person.ID was not set) + more

+++ src/main/java/models/Produksjonsselskap.java
@@ -16,7 +16,7 @@ public class Produksjonsselskap implements ActiveDomainObject {
     private String navn;
     private Date opprettet;
 
-    public Produksjonsselskap(int id){
+    public Produksjonsselskap(long id){
         this.ID = id;
     }
